Ignore compiler warnings in Boost headers
Because Boost is a 3rd-party library, it's not really up to us to silence
every last warning emitted by the compiler about it. The best way to
prevent the compiler from doing that is to express that those headers are
'system' headers, i.e. external headers found using the `-isystem` compiler
flag instead of `-I`. CMake lets you do this with the `SYSTEM` flag.